Moved helper macro methods to seperate trait.
